Correct profile name in inspec.yml
Back during the PR to go from 2.6 to 3.6, the name: in the inspec.yml was incorrectly changed from redhat-enterprise-linux-7-stig-baseline to RHEL_7_STIG. This puts it back, using the convention that uses our repo-naming standard. This is important for clarity but more importantly to allow overlay profiles that have depends statements work correctly.